Fix use of interpretTimingParam()
This fixes a bug where the timing argument to the constructor of
(Par|Seq)AnimationGroupTemplate was not correctly converted to a Timing object.
As a result, default values for timing properties were not applied.

Also remove superfluous explicit setting of timing.fill for DEFAUKT_GROUP, as
this is now the default.