Fix use of isDefinedAndNotNull()
When processing arguments to API functions, we generally want to test only
whether a variable is defined. We then rely on the regular JavaScript type
conversion rules for handling values of the wrong type, including null.
Introduce a new isDefined() helper for this case.

Also fix the test for a parent in TimedItem(). This was broken in
2998dbbd9c35c748cba75ee9990a621ae60f2fa9, which caused test-parent to fail.