Fix calculation of TimedItem.itemTime
When calculating the item time, use the parent iteration time when it's
non-null, even if it's zero. See
https://dvcs.w3.org/hg/FXTF/raw-file/tip/web-anim/index.html#calculating-the-item-time

When handling fill mode, use an inclusive bound at the start of the animation,
as we do at the end of the animation.

This fixes the behaviour of fill modes of child animations in iterating parent
groups. Note that the expected values for some tests were incorrect and have
been updated.