Calculate TimedItem.endTime lazily
When the timing properties of a TimedItem in a hierarchy are modified, the
endTime must be updated before the parent is updated (in the case of a parallel
parent), but the endTime may also be affected by that update to the parent (in
the case of a sequence parent). To avoid this complexity, we calculate the
endTime lazily in the property getter.

This will also allow us to reduce the number of calls to updateTimeMarkers when
the timing properties of a TimedItem in a hierarchy are modified.