Fix TimedItem._parentToGlobalTime() to use DEFAULT_GROUP.itemTime
Currently this method uses DEFAULT_GROUP.currentTime, but this is always zero
due to bug https://github.com/web-animations/web-animations-js/issues/66. This
bug probably won't get fixed, as the switch to TimeSources is imminent.

As a work-around, this change fixes TimedItem._parentToGlobalTime() to use
DEFAULT_GROUP.itemTime, which is equivalent.