Rename TimedItem._getActiveAnimations() to _getItemsInEffect()
This method gets all TimedItems which are currently in effect. This is a
superset of the TimedItems in their active interval, as a TimedItem can have an
effect outside its active interval due to fill.

Also add a stub implementation of this method to TimedItem, since
AnimationGroup._getItemsInEffect() calls _getItemsInEffect() on its children,
which are of type TimedItem.