Use inclusive bounds for a TimedItem's animation interval
Currently, the animation interval of a TimedItem uses exclusive bounds. So fill
must be used to make the animation take effect at these bounds.

This change modifies TimedItem to use inclusive bounds, updates existing tests
and adds a new test. This fixes
https://github.com/web-animations/web-animations-js/issues/80

Note that change b20eaa75 modified the start bound of the animation interval to
be exclusive. This was done as part of a change to fix the handling of an item
time of zero. It seems that the start bound was made exclusive to match the end
bound and to minimise the change in behaviour.