Fix fromCssValue() methods to return undefined if the input value is invalid
Currently, the various fromCssValue() methods handle bad input differently.
This change standardises the handling so that missing logic is more obvious.

It also adds a console.assert() to the main fromCssValue() function to detect
this condition, and a TODO to fix the keyframe normalization process to ensure
the assert is never hit.