Adding support for expected failure under some platforms.
This is done by adding an "expected_failures" object before the bootstrap.js
file is loaded.

The structrue looks like;
var expected_failures = {
 'Test 1': {
   chrome: ["28.0"],
 },
 'Test 2': {
   firefox: true,
 },
 'Test 3': {
   c: true,
 },
};

Each test() or timing_test() can be expected to fail as long as the name can be
specified.

This uses bowser (https://github.com/ded/bowser) for browser detection and any
supported bowser value can be added in the expected failures. For example,
 * Browser name, "firefox", "chrome"
 * Browser engine, "webkit", "gecko"
 * Browser basic version, "28.0"
 * Yahoo's Graded Browser support, "a", "b", "c"